Move MQTT dev tools to integrations (#6189)
