+++ src/panels/config/devices/device-detail/integration-elements/mqtt/mqtt-discovery-payload.ts
