+++ src/panels/config/devices/device-detail/integration-elements/mqtt/mqtt-messages.ts
@@ -9,9 +9,9 @@ import {
   TemplateResult,
 } from "lit-element";
 import { classMap } from "lit-html/directives/class-map";
-import { formatTimeWithSeconds } from "../../common/datetime/format_time";
-import { HomeAssistant } from "../../types";
-import { MQTTMessage } from "../../data/mqtt";
+import { formatTimeWithSeconds } from "../../../../../../common/datetime/format_time";
+import { HomeAssistant } from "../../../../../../types";
+import { MQTTMessage } from "../../../../../../data/mqtt";
 
 @customElement("mqtt-messages")
 class MQTTMessages extends LitElement {
