+++ src/panels/developer-tools/ha-panel-developer-tools.ts
@@ -13,7 +13,6 @@ import {
   property,
   TemplateResult,
 } from "lit-element";
-import { isComponentLoaded } from "../../common/config/is_component_loaded";
 import scrollToTarget from "../../common/dom/scroll-to-target";
 import { navigate } from "../../common/navigate";
 import "../../components/ha-menu-button";
@@ -72,15 +71,6 @@ class PanelDeveloperTools extends LitElement {
                 "ui.panel.developer-tools.tabs.events.title"
               )}
             </paper-tab>
-            ${isComponentLoaded(this.hass, "mqtt")
-              ? html`
-                  <paper-tab page-name="mqtt">
-                    ${this.hass.localize(
-                      "ui.panel.developer-tools.tabs.mqtt.title"
-                    )}
-                  </paper-tab>
-                `
-              : ""}
           </paper-tabs>
         </app-header>
         <developer-tools-router
