+++ src/translations/en.json
@@ -1506,6 +1506,20 @@
             "create": "Create"
           }
         },
+        "mqtt": {
+          "button": "Configure",
+          "title": "MQTT",
+          "description_publish": "Publish a packet",
+          "topic": "topic",
+          "payload": "Payload (template allowed)",
+          "publish": "Publish",
+          "description_listen": "Listen to a topic",
+          "listening_to": "Listening to",
+          "subscribe_to": "Topic to subscribe to",
+          "start_listening": "Start listening",
+          "stop_listening": "Stop listening",
+          "message_received": "Message {id} received on {topic} at {time}:"
+        },
         "zha": {
           "button": "Configure",
           "header": "Configure Zigbee Home Automation",
@@ -2298,19 +2312,6 @@
             "stop_listening": "Stop listening",
             "alert_event_type": "Event type is a mandatory field",
             "notification_event_fired": "Event {type} successful fired!"
-          },
-          "mqtt": {
-            "title": "MQTT",
-            "description_publish": "Publish a packet",
-            "topic": "topic",
-            "payload": "Payload (template allowed)",
-            "publish": "Publish",
-            "description_listen": "Listen to a topic",
-            "listening_to": "Listening to",
-            "subscribe_to": "Topic to subscribe to",
-            "start_listening": "Start listening",
-            "stop_listening": "Stop listening",
-            "message_received": "Message {id} received on {topic} at {time}:"
           },
           "services": {
             "title": "Services",
